Making sure the winsock library is initialized from the main executable only (not from any DLLMain routine)
- this is a Windows specific change
